Alpha_2.0.0 scrolling capabilities
- refactored entire project with new title: TerrainRenderer
- package is now com.example.terrainrenderer
- renders flat ground with green for 1s and grey for 0s
- buttons for scrolling left and right
Operation:
1. Install and run the app. It should display a single block on the
bottom left corner of the screen and a message telling you that it can't
find DemoTerrain.csv. This creates the folder path to the file, but not
the file itself.
2. Go into Android/data/com.example.terrainrenderer/Terrain and put
DemoTerrain.csv there. Note: you don't enjoy the scrolling unless your
file is big enough to scroll through, so make it have at least 20 ones
and zeros in it.
3. Rerun the app, be amazed.